# swagger.json
{
    "swagger": "2.0",
    "basePath": "\/",
    "info": {
        "title": "API Escape Game Corp.",
        "version": "1.0.0",
        "description": "Service web pour gerer les informations de EGC"
    },
    "paths": {
...
        "\/parties": {
            "get": {
                "tags": [
                    "Parties"
                ],
                "operationId": "getPartiesCollection",
                "produces": [
                    "application\/json",
                    "text\/html"
                ],
                "summary": "Retrieves the collection of Parties resources.",
                "responses": {
                    "200": {
                        "description": "Parties collection response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/Parties"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "The collection page number",
                        "type": "integer"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parties"
                ],
                "operationId": "postPartiesCollection",
                "consumes": [
                    "application\/json",
                    "text\/html"
                ],
                "produces": [
                    "application\/json",
                    "text\/html"
                ],
                "summary": "Creates a Parties resource.",
                "responses": {
                    "201": {
                        "description": "Parties resource created",
                        "schema": {
                            "$ref": "#\/definitions\/Parties"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "parameters": [
                    {
                        "name": "parties",
                        "in": "body",
                        "description": "The new Parties resource",
                        "schema": {
                            "$ref": "#\/definitions\/Parties"
                        }
                    }
                ]
            }
        },
        "\/parties\/byDate\/{date}": {
            "get": {
                "operationId": "getByDate",
                "parameters": [
                    {
                        "name": "date",
                        "required": true,
                        "type": "string",
                        "in": "path",
                        "description": "Affiche les parties en fonction de la date"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "The collection page number",
                        "type": "integer"
                    }
                ],
                "produces": [
                    "application\/json"
                ],
                "tags": [
                    "Parties"
                ],
                "summary": "Retrieves the collection of Parties resources.",
                "responses": {
                    "200": {
                        "description": "Parties collection response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/Parties"
                            }
                        }
                    }
                }
            }
        },
        "\/parties\/byDate\/{date}\/{heure}": {
            "get": {
                "operationId": "getByDateEtHeure",
                "parameters": [
                    {
                        "name": "date",
                        "required": true,
                        "type": "string",
                        "in": "path",
                        "description": "Date de la partie"
                    },
                    {
                        "name": "heure",
                        "required": true,
                        "type": "string",
                        "in": "path",
                        "description": "Heure de la partie"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "The collection page number",
                        "type": "integer"
                    }
                ],
                "produces": [
                    "application\/json"
                ],
                "tags": [
                    "Parties"
                ],
                "summary": "Retrieves the collection of Parties resources.",
                "responses": {
                    "200": {
                        "description": "Parties collection response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/Parties"
                            }
                        }
                    }
                }
            }
        },
        "\/parties\/byJoueur\/{joueur}": {
            "get": {
                "operationId": "getByJoueur",
                "parameters": [
                    {
                        "name": "joueur",
                        "required": true,
                        "type": "string",
                        "in": "path",
                        "description": "ID du joueur"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "The collection page number",
                        "type": "integer"
                    }
                ],
                "produces": [
                    "application\/json"
                ],
                "tags": [
                    "Parties"
                ],
                "summary": "Retrieves the collection of Parties resources.",
                "responses": {
                    "200": {
                        "description": "Parties collection response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/Parties"
                            }
                        }
                    }
                }
            }
        },
        "\/parties\/{id}": {
            "get": {
                "tags": [
                    "Parties"
                ],
                "operationId": "getPartiesItem",
                "produces": [
                    "application\/json",
                    "text\/html"
                ],
                "summary": "Retrieves a Parties resource.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parties resource response",
                        "schema": {
                            "$ref": "#\/definitions\/Parties"
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Parties"
                ],
                "operationId": "putPartiesItem",
                "consumes": [
                    "application\/json",
                    "text\/html"
                ],
                "produces": [
                    "application\/json",
                    "text\/html"
                ],
                "summary": "Replaces the Parties resource.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "parties",
                        "in": "body",
                        "description": "The updated Parties resource",
                        "schema": {
                            "$ref": "#\/definitions\/Parties"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parties resource updated",
                        "schema": {
                            "$ref": "#\/definitions\/Parties"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Parties"
                ],
                "operationId": "deletePartiesItem",
                "summary": "Removes the Parties resource.",
                "responses": {
                    "204": {
                        "description": "Parties resource deleted"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "in": "header",
            "description": "Value for the Authorization header",
            "name": "Authorization"
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ],
...
}